Add cohort selection, SteamCharts scraper, and database loading pipeline
- KaggleLoader: switched from corrupted CSV to JSON export, added
  select_cohort() for reproducible cohort selection
- SteamChartsScraper: scrapes historical monthly player data with
  caching, rate limiting, and batch failure handling
- Fixed avg_players schema type (INTEGER -> REAL) to correctly
  store decimal averages
- Loaded 195-game cohort (2019-2022 releases, 500+ reviews, with
  active SteamCharts history) into steam_project.db
- Added exploration notebook covering data validation and cohort
  selection

+++ schema.sql
@@ -63,7 +63,7 @@ CREATE TABLE game_genres (
 CREATE TABLE monthly_metrics (
     appid           INTEGER NOT NULL,
     month           DATE NOT NULL,      -- store as first-of-month, e.g. '2021-03-01'
-    avg_players     INTEGER,
+    avg_players     REAL,
     peak_players    INTEGER,
     est_owners_low  INTEGER,
     est_owners_high INTEGER,
